minitap/ai_agent_menu/6_mcp_toolbox/requirements.txt
# MiniTAP: MCP Toolboxを活用したAgentic Data Analytics入門
# 必要なPythonライブラリ

# MCP Toolbox Core SDK
toolbox-core>=0.5.0

# Google Cloud プラットフォーム
google-cloud-aiplatform==1.97.0
google-adk==1.4.1
google-genai==1.20.0
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
db-dtypes

# データ処理・分析
pandas
numpy
nest-asyncio

# 可視化
matplotlib
seaborn
plotly
japanize-matplotlib

# Jupyter Notebook サポート
jupyter
ipython

# HTTP クライアント
requests
aiohttp

# ユーティリティ
pydantic
python-dotenv
structlog

# セキュリティ
cryptography